+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [1.0.1] - Unreleased
+## [1.0.1] - 2021-12-20
 ### Added
 - Add public functions for connecting (connectToAstarte) and disconnecting (disconnectFromAstarte).
 - Add support for all array types by adding QList<T> overloads to sendData